requirements.txt
# Core dependencies
numpy==1.24.4  # Compatible with Python 3.9 and other packages
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
python-dotenv>=1.0.0
Pillow>=10.0.0
imagededup>=0.3.0
scikit-image>=0.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
ImageHash>=4.3.0
transformers>=4.30.0
timm>=0.9.0

# PyTorch - CPU version for compatibility
torch==2.2.2
torchvision==0.17.2

# Image Processing
rawpy>=0.18.0
exifread>=3.0.0

# UI and Progress
rich>=13.0.0
colorama>=0.4.6
tqdm>=4.65.0

# System and Utility
psutil>=5.9.0
certifi>=2023.0.0
python-dateutil>=2.8.2
requests>=2.31.0
gdown>=4.7.0
google-generativeai>=0.3.0
ultralytics>=8.0.0
facenet_pytorch>=2.5.0

# Dependency version constraints
protobuf>=4.25.3,<5.0
grpcio-status>=1.58.0,<2.0
mediapipe>=0.10.0,<0.11.0

# Additional dependencies for proper functioning
matplotlib>=3.5.0
pandas>=1.5.0
